% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Age_function.R
\name{Age_function}
\alias{Age_function}
\title{An Age function}
\usage{
Age_function(year)
}
\arguments{
\item{year}{which year do you want the player's name for?}
}
\description{
This function allows you to enter a year and extract the name of the oldest player
}
\examples{
Age_function()
}
