% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corr_matrix.R
\name{get_corr_matrix}
\alias{get_corr_matrix}
\title{Correlation Matrix for All Numeric Variables in a Given Year}
\usage{
get_corr_matrix(year)
}
\arguments{
\item{year}{The year (season) for which you are interested in getting a correlation matrix.}
}
\description{
This function allows you to output the correlation matrix for all numeric variables in a given year.
}
\examples{
get_corr_matrix()
}
\keyword{correlation,}
\keyword{matrix}
\keyword{year,}
