% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_minutes_played.R
\name{get_minutes_played}
\alias{get_minutes_played}
\title{Calculating Players' Total Minutes Played for a Given Year}
\usage{
get_minutes_played(year)
}
\arguments{
\item{year}{The year (season) for which you are interested in knowing the players' total minutes played.}
}
\description{
This function allows you to calculate the total minutes played for a given year for all players.
}
\examples{
get_minutes_played()
}
\keyword{minutes,}
\keyword{mp,}
\keyword{played}
\keyword{year,}
