% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oldest_player.R
\name{get_oldest_player}
\alias{get_oldest_player}
\title{Finding the Oldest Player for a Given Year}
\usage{
get_oldest_player(year)
}
\arguments{
\item{year}{The year (season) for which you are interested in knowing the oldest player.}
}
\description{
This function allows you to find the oldest player in a given year. For ties, it returns the first player in alphabetical order.
}
\examples{
get_oldest_player()
}
\keyword{age,}
\keyword{player}
\keyword{year,}
