% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Numeric_Only.R
\name{Numeric_Only}
\alias{Numeric_Only}
\title{A numeric function}
\usage{
Numeric_Only(year)
}
\arguments{
\item{year}{which year do you want the player's name for?}
}
\description{
This function allows you to enter a year and filter using that year and creates a correlation matrix only numeric variables
}
\examples{
Numeric_Only()
}
