% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Year_function.R
\name{Year_function}
\alias{Year_function}
\title{A Year function}
\usage{
Year_function(year)
}
\arguments{
\item{year}{which year do you want the player's name for?}
}
\description{
This function allows you to enter a year and extract the name of the player with the highest Free throw percentage for that year.
}
\examples{
Year_function()
}
\keyword{max}
